moved getTip out to it's own file
