+++ app/src/lib/git/branch.ts
@@ -1,11 +1,7 @@
 import { git, envForAuthentication } from './core'
 import { Repository } from '../../models/repository'
 import { Branch, BranchType } from '../../models/branch'
-import { Tip, TipState } from '../../models/tip'
 import { User } from '../../models/user'
-import { getCurrentBranch } from './for-each-ref'
-
-import { fatalError } from '../../lib/fatal-error'
 
 /** Create a new branch from the given start point. */
 export async function createBranch(repository: Repository, name: string, startPoint: string): Promise<void> {
@@ -35,38 +31,4 @@ export async function deleteBranch(repository: Repository, branch: Branch, user:
   }
 
   return true
-}
-
-/** Get the name of the current branch. */
-export async function getTip(repository: Repository): Promise<Tip> {
-
-  const revParse = await git([ 'rev-parse', 'HEAD' ], repository.path, 'getTip', { successExitCodes: new Set([ 0, 128 ]) })
-  if (revParse.exitCode === 128) {
-    // fatal: ambiguous argument 'HEAD': unknown revision or path not in the working tree.
-    return {
-      kind: TipState.Unborn,
-    }
-  }
-
-  const currentSha = revParse.stdout.trim()
-
-  const symbolicRef = await git([ 'symbolic-ref', 'HEAD' ], repository.path, 'getTip', { successExitCodes: new Set([ 0, 128 ]) })
-  if (symbolicRef.exitCode === 128) {
-    // fatal: ref HEAD is not a symbolic ref
-    return {
-      kind: TipState.Detached,
-      currentSha,
-    }
-  }
-
-  const currentBranch = await getCurrentBranch(repository)
-  if (!currentBranch) {
-    fatalError(`getTip failed despite all the previous guard checks`)
-    return { kind: TipState.Unknown }
-  }
-
-  return {
-    kind: TipState.Valid,
-    branch: currentBranch,
-  }
 }
