+++ app/src/lib/git/index.ts
@@ -20,6 +20,7 @@ export * from './reset'
 export * from './rev-list'
 export * from './rev-parse'
 export * from './status'
+export * from './tip'
 export * from './update-ref'
 export * from './var'
 export * from './merge'
